Revert "1대다 채팅방 완전구현"
This reverts commit 64796705296848d25fb8911fe2d74d7e35ca82b8.
