+++ src/main/java/com/catchbug/biz/admin/stock/StockController.java
@@ -3,13 +3,13 @@ package com.catchbug.biz.admin.stock;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.catchbug.biz.vo.CategoryVO;
+import com.catchbug.biz.vo.MemberVO;
 import com.catchbug.biz.vo.ProductVO;
 
 @Controller
@@ -20,21 +20,7 @@ public class StockController {
 
 	// 상품 리스트
 	@RequestMapping("/stockList.do")
-	public String StockList(Model model,CategoryVO vo) {
-	// 처음 들어갔을때 카테고리 불러오기
-
-		
-		// 대분류 카테고리
-		model.addAttribute("mainCategory",ss.getMainCategory());
-		//소분류 카테고리
-		List<CategoryVO> category = ss.getSubCategory();
-		model.addAttribute("subCategory",category);
-		// 첫 요청 상품 데이터 최근 등록순
-		vo.setSub_category(0);
-		model.addAttribute("product", ss.getProductList(vo));
-		
-		System.out.println(ss.getSubCategory());
-		System.out.println(ss.getMainCategory());
+	public String StockList() {
 		return "admin/stock_list";
 	}
 
@@ -79,7 +65,6 @@ public class StockController {
 //	}
 	
 	// 하위 카테고리 클릭시 상품 비동기처리
-	
 	@RequestMapping("/orderAjax.do")
 	@ResponseBody
 	public List<ProductVO> orderAjax(CategoryVO vo){
