+++ src/main/java/com/catchbug/biz/chat/ChatController.java
@@ -1,4 +1,4 @@
-package com.catchbug.biz.board.chat;
+package com.catchbug.biz.chat;
 
 import javax.servlet.http.HttpSession;
 
@@ -6,15 +6,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import com.catchbug.biz.vo.MemberVO;
-
 @Controller
 public class ChatController {
 
 	@GetMapping("/chat")
-	public void chat(Model model,HttpSession session) {
-		MemberVO user = (MemberVO) session.getAttribute("memberId.ceo");
-		model.addAttribute("userid", user);
-	}
+	public String chatGET() {
 
+		System.out.println("@ChatController, chat GET()");
+
+		return "chat";
+	}
 }
