+++ src/main/java/com/catchbug/biz/vo/ProductVO.java
@@ -2,8 +2,6 @@ package com.catchbug.biz.vo;
 
 import java.util.Date;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-
 import lombok.Data;
 
 @Data
@@ -14,7 +12,6 @@ public class ProductVO {
 	private String product_name;
 	private String product_img;
 	private String brand;
-	@JsonFormat(shape=JsonFormat.Shape.STRING, pattern="yyyy-MM-dd")
 	private Date add_day;
 	private String uuid;
 }
