+++ src/main/webapp/WEB-INF/views/board/chat.jsp
@@ -112,7 +112,7 @@
                 send();
             });
 
-            /* const websocket = new WebSocket("ws://192.168.0.39:8080/ws/chat"); */
+            const websocket = new WebSocket("ws://192.168.0.39:8080/ws/chat");
 
             websocket.onmessage = onMessage;
             websocket.onopen = onOpen;
