--- src/main/webapp/WEB-INF/views/board/chatBoard.jsp
@@ -1,214 +1,0 @@
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
-
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-<meta charset="utf-8">
-<meta content="width=device-width, initial-scale=1.0" name="viewport">
-
-<title>게시판 - 일대일 상담</title>
-<meta content="" name="description">
-<meta content="" name="keywords">
-
-<!-- Favicons -->
-<link href="assets/img/favicon.png" rel="icon">
-<link href="assets/img/apple-touch-icon.png" rel="apple-touch-icon">
-
-<!-- Google Fonts -->
-<link href="https://fonts.gstatic.com" rel="preconnect">
-<link
-	href="https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i,700,700i|Nunito:300,300i,400,400i,600,600i,700,700i|Poppins:300,300i,400,400i,500,500i,600,600i,700,700i"
-	rel="stylesheet">
-
-<!-- Vendor CSS Files -->
-<link href="assets/vendor/bootstrap/css/bootstrap.min.css"
-	rel="stylesheet">
-<link href="assets/vendor/bootstrap-icons/bootstrap-icons.css"
-	rel="stylesheet">
-<link href="assets/vendor/boxicons/css/boxicons.min.css"
-	rel="stylesheet">
-<link href="assets/vendor/quill/quill.snow.css" rel="stylesheet">
-<link href="assets/vendor/quill/quill.bubble.css" rel="stylesheet">
-<link href="assets/vendor/remixicon/remixicon.css" rel="stylesheet">
-<link href="assets/vendor/simple-datatables/style.css" rel="stylesheet">
-
-<!-- Template Main CSS File -->
-<link href="assets/css/style.css" rel="stylesheet">
-
-<!-- Jquery 선언 -->
-<script src="https://code.jquery.com/jquery-3.6.0.min.js"
-	integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4="
-	crossorigin="anonymous"></script>
-
-<!-- 채팅페이지 디자인 -->
-<script
-	src="https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js"></script>
-
-<!-- =======================================================
-  * Template Name: NiceAdmin - v2.3.1
-  * Template URL: https://bootstrapmade.com/nice-admin-bootstrap-admin-html-template/
-  * Author: BootstrapMade.com
-  * License: https://bootstrapmade.com/license/
-  ======================================================== -->
-
-<script
-	src="https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js"></script>
-</head>
-<body>
-
-	<!-- ======= Header ======= -->
-
-	<jsp:include page="../mainInclude/header.jsp"></jsp:include>
-
-	<!-- End Header -->
-
-
-
-
-	<!-- ======= Sidebar ======= -->
-
-	<jsp:include page="../mainInclude/sidebar.jsp"></jsp:include>
-
-	<!-- ======= Sidebar ======= -->
-
-	<main id="main" class="main">
-	<div class="container">
-		<div class="col-6">
-			<label><b>채팅방</b></label>
-		</div>
-		<div>
-			<div id="msgArea" class="col"></div>
-			<div class="col-6">
-				<div class="input-group mb-3">
-					<input type="text" id="msg" class="form-control"
-						aria-label="Recipient's username" aria-describedby="button-addon2">
-					<div class="input-group-append">
-						<button class="btn btn-outline-secondary" type="button"
-							id="button-send">전송</button>
-					</div>
-				</div>
-			</div>
-		</div>
-		<div class="col-6"></div>
-	</div>
-
-	</main>
-
-
-	<!--  카드바디 -->
-
-	<!-- 본문 -->
-
-
-	<!-- ======= Footer ======= -->
-	<jsp:include page="../mainInclude/footer.jsp"></jsp:include>
-	<!-- End Footer -->
-	<a href="#"
-		class="back-to-top d-flex align-items-center justify-content-center"><i
-		class="bi bi-arrow-up-short"></i></a>
-	<!-- Vendor JS Files -->
-	<script src="assets/vendor/apexcharts/apexcharts.min.js"></script>
-	<script src="assets/vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
-	<script src="assets/vendor/chart.js/chart.min.js"></script>
-	<script src="assets/vendor/echarts/echarts.min.js"></script>
-	<script src="assets/vendor/quill/quill.min.js"></script>
-	<script src="assets/vendor/simple-datatables/simple-datatables.js"></script>
-	<script src="assets/vendor/tinymce/tinymce.min.js"></script>
-	<script src="assets/vendor/php-email-form/validate.js"></script>
-	<!-- Template Main JS File -->
-	<script src="assets/js/main.js"></script>
-
-	<!-- 채팅 이용을 위한 자바스크립트 코드-->
-	<script src="https://cdn.jsdelivr.net/sockjs/1/sockjs.min.js"></script>
-	<script type="text/javascript">
-		$(document).ready(function() {
-
-			var sock = new SockJS('http://localhost:8080/chat');
-			var username = '${memberId.ceo}';
-
-            sock.onmessage = onMessage;
-            sock.onopen = onOpen;
-            sock.onclose = onClose;
-
-			$("#disconn").on("click", function(e){
-				disconnect();
-			});
-
-			//전송 버튼 누르는 이벤트
-			$("#button-send").on("click", function(e){
-				send();
-				$('#msg').val('')
-			});
-
-			//메세지 전송 스크립트
-			function send() {
-
-				let msg = $('#msg').val();
-				
-
-				console.log(username + "내가보낸값 찍어보자" + msg);
-				sock.send(username + ":" + msg);
-				msg.value = '';
-			}
-
-			//채팅창에서 나갔을 때
-			function onClose(evt) {
-				var str = username + ": 님이 방을 나가셨습니다.";
-				sock.send(str);
-			}
-			//채팅창에 들어왔을 때
-			function onOpen(evt) {
-				var str = username + ": 님이 입장하셨습니다.";
-				sock.send(str);
-			}
-
-			function onMessage(msg) {
-				var data = msg.data;
-				var sessionId = null;
-				//데이터를 보낸 사람
-				var message = null;
-				var arr = data.split(":");
-				console.log(msg);
-
-				for (var i = 0; i < arr.length; i++) {
-					console.log('배열arr[' + i + ']: ' + arr[i]);
-				}
-
-				var cur_session = username;
-
-				//현재 세션에 로그인 한 사람
-				sessionId = arr[0];
-				message = arr[1];
-
-				console.log("sessionID : " + sessionId);
-				console.log("로그인한 아이디 : " + cur_session);
-
-				//로그인 한 클라이언트와 타 클라이언트를 분류하기 위함
-				if (sessionId == cur_session) {
-					var str = "<div class='col-6'>";
-					str += "<div class='alert alert-secondary'>";
-					str += "<b>" + sessionId + " : " + message + "</b>";
-					str += "</div></div>";
-					$("#msgArea").append(str);
-				} else {
-					var str = "<div class='col-6'>";
-					str += "<div class='alert alert-warning'>";
-					str += "<b>" + sessionId + " : " + message + "</b>";
-					str += "</div></div>";
-					$("#msgArea").append(str);
-				}
-			}
-		})
-	</script>
-
-	<!-- bootstrap function -->
-	<script type="text/javascript">
-		$(function() {
-			$('[data-toggle="tooltip"]').tooltip();
-		});
-	</script>
-</body>
-</html>
