+++ src/main/webapp/WEB-INF/views/mainInclude/header.jsp
@@ -12,6 +12,8 @@
 
 
 <!-- ======= Header ======= -->
+=======
+>>>>>>> d24ae394f56ad68c6f2fd2ce7cafb4f8c9e20882
 
 <script src="https://code.jquery.com/jquery-1.12.4.min.js"></script>
 
@@ -142,5 +144,88 @@
 
 	</nav>
 	<!-- info drop down end -->
+
+
+	<script
+		src="https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js"></script>
+	<script>
+            $(document).ready(function(){
+
+            const username = '${memberId.ceo}'
+
+            $("#disconn").on("click", (e) => {
+                disconnect();
+            })
+            
+            $("#button-send").on("click", (e) => {
+                send();
+            });
+
+            const websocket = new WebSocket("ws://192.168.0.39:8080/ws/chat");
+
+            websocket.onmessage = onMessage;
+            websocket.onopen = onOpen;
+            websocket.onclose = onClose;
+
+            function send(){
+
+                let msg = document.getElementById("msg");
+
+                console.log(username + ">>" + msg.value);
+                websocket.send(username + ">>" + msg.value);
+                msg.value = '';
+            }
+            
+            //채팅창에서 나갔을 때
+            function onClose(evt) {
+                var str = username + ": 님이 방을 나가셨습니다.";
+                websocket.send(str);
+            }
+            
+            //채팅창에 들어왔을 때
+            function onOpen(evt) {
+                var str = username + ": 님이 입장하셨습니다.";
+                websocket.send(str);
+            }
+
+            function onMessage(msg) {
+                var data = msg.data;
+                var sessionId = null;
+                //데이터를 보낸 사람
+                var message = null;
+                var arr = data.split(":");
+
+                for(var i=0; i<arr.length; i++){
+                    console.log('arr[' + i + ']: ' + arr[i]);
+                }
+
+                var cur_session = username;
+
+                //현재 세션에 로그인 한 사람
+                console.log("cur_session : " + cur_session);
+                sessionId = arr[0];
+                message = arr[1];
+
+                console.log("sessionID : " + sessionId);
+                console.log("cur_session : " + cur_session);
+
+                //로그인 한 클라이언트와 타 클라이언트를 분류하기 위함
+                if(sessionId == cur_session){
+                    var str = "<div class='col-6'>";
+                    str += "<div class='alert alert-secondary'>";
+                    str += "<b>" + sessionId + " : " + message + "</b>";
+                    str += "</div></div>";
+                    $("#msgArea").append(str);
+                }
+                else{
+                    var str = "<div class='col-6'>";
+                    str += "<div class='alert alert-warning'>";
+                    str += "<b>" + sessionId + " : " + message + "</b>";
+                    str += "</div></div>";
+                    $("#msgArea").append(str);
+                }
+            }
+            })
+</script>
 
 </header>
